Make the default disk be /dev/shm/vhostmd0.
Previously we used a subdirectory (/dev/shm/vhostmd/).  However
the permissions on the subdirectory had to be set by hand by the
system administrator, and if the sysadmin didn't set them, then
libvirt would fail to start the guest.

However by just putting the metrics disk into /dev/shm directly,
the permissions can be controlled by vhostmd when it creates the
device.

Not having the subdirectory is also simpler.