Add CONNECT substitution <action>s.
Allow CONNECT to be substituted in any <action>s in order to
pass the correct libvirt URI down to actions such as virsh.  The
correct way to run virsh is now:

  virsh -r CONNECT cmd [...]

which substitutes as either:

  virsh -r --connect 'uri' cmd [...]

or:

  virsh -r cmd [...]

Note that you should always run virsh readonly (-r option).

Also corrects the documentation: ID should be VMID.

Also this cleans up the string replacement code and ensures that
it doesn't leak memory on failure.