Make Xen Libraries really optional
The default behavior in the configure script was to make with_xenstore
set to 'true', which caused the compilation to break in my case because
I don't have xen libraries, therefore, xs.h

Now the default behavior is set according to its existence or not. Same
thing for xenstat and xenctrl.

Signed-off-by: Eduardo Otubo <eduardo.otubo@gmail.com>