Change actions to retrieve vendor and product info
On a KVM host, the action to report VirtualizationVendor checks the
vendor of the libvirt RPM, which is typically a meta-package that pulls
in all of libvirt. This is undesirable on a minimalistic hypervisor.
Change the action to check the vendor of the package owning
/etc/os-release, which should be available on even the most minimal
systems and works for both Xen and KVM.

The VirtualizationProductInfo action reports inconsistent information
between Xen and KVM. On a Xen host the Xen version is reported, whereas
the libvirt version is reported on a KVM host. Change the action to
report the "Running hypervisor" as reported by 'virsh version'.

Fixes: https://github.com/vhostmd/vhostmd/issues/7

Signed-off-by: Jim Fehlig <jfehlig@suse.com>